+++ src/routes/router.tsx
@@ -4,7 +4,7 @@ import SignIn from '@pages/auth/SignIn';
 import SignUp from '@pages/auth/SignUp';
 import App from '@/App';
 import Recruit from '@/pages/recruit';
-import VisaRegistration from '@/pages/employee/visaRegistration';
+import VisaRegistration from '@/pages/visaRegistration';
 import PostNotice from '@/pages/employer/postNotice/PostNotice';
 import Home from '@/pages/home';
 import ApplyGuide from '@/pages/apply/applyguide/ApplyGuide';
@@ -28,7 +28,7 @@ export const router = createBrowserRouter([
       { path: ROUTE_PATH.APPLY.GUIDE, element: <ApplyGuide /> },
       { path: ROUTE_PATH.APPLY.APPLYPAGE, element: <ApplyPage /> },
       { path: ROUTE_PATH.RECRUIT, element: <Recruit /> },
-      { path: ROUTE_PATH.EMPLOYEE.VISA_REGISTRATION, element: <VisaRegistration /> },
+      { path: ROUTE_PATH.VISA_REGISTRATION, element: <VisaRegistration /> },
       { path: ROUTE_PATH.POST_NOTICE, element: <PostNotice /> },
       { path: ROUTE_PATH.COMPANY_RECRUITMENTS, element: <CompanyRecruitments /> },
       { path: ROUTE_PATH.APPLICANTS, element: <Applicants /> },
